Refactor/jpa mybatis -> jpa 전환

See merge request re12io/happyhouse5!7

+++ src/main/java/com/ssafy/happyhouse5/service/HouseService.java
@@ -1,12 +1,13 @@
 package com.ssafy.happyhouse5.service;
 
-import com.ssafy.happyhouse5.dto.house.HouseDeal;
-import com.ssafy.happyhouse5.dto.house.HouseInfo;
+
+import com.ssafy.happyhouse5.entity.HouseDeal;
+import com.ssafy.happyhouse5.entity.HouseInfo;
 import java.util.List;
 
 public interface HouseService {
 
     List<HouseInfo> findHouseInfoByDongCode(String dongCode);
 
-    List<HouseDeal> findHouseDealListByAptCode(int aptCode);
+    List<HouseDeal> findHouseDealListByAptCode(Long aptCode);
 }
